refactor: use gtc storage driver
